Extract monkey patched artefact logic to a module
It is much safer to do monkey patching in a module if it really has to be done.
